+++ NCCommunication/NCCommunicationBackground.swift
@@ -123,7 +123,7 @@ import Foundation
         guard let url = downloadTask.currentRequest?.url?.absoluteString.removingPercentEncoding else { return }
         let fileName = (url as NSString).lastPathComponent
         let serverUrl = url.replacingOccurrences(of: "/"+fileName, with: "")
-        let progress = Double(Double(totalBytesWritten)/Double(totalBytesExpectedToWrite))
+        let progress = totalBytesWritten/totalBytesExpectedToWrite
 
         DispatchQueue.main.async {
             NCCommunicationCommon.shared.delegate?.downloadProgress?(progress, totalBytes: totalBytesWritten, totalBytesExpected: totalBytesExpectedToWrite, fileName: fileName, serverUrl: serverUrl, session: session, task: downloadTask)
@@ -153,7 +153,7 @@ import Foundation
         guard let url = task.currentRequest?.url?.absoluteString.removingPercentEncoding else { return }
         let fileName = (url as NSString).lastPathComponent
         let serverUrl = url.replacingOccurrences(of: "/"+fileName, with: "")
-        let progress = Double(Double(totalBytesSent)/Double(totalBytesExpectedToSend))
+        let progress = totalBytesSent/totalBytesExpectedToSend
 
         DispatchQueue.main.async {
             NCCommunicationCommon.shared.delegate?.uploadProgress?(progress, totalBytes: totalBytesSent, totalBytesExpected: totalBytesExpectedToSend, fileName: fileName, serverUrl: serverUrl, session: session, task: task)
@@ -162,7 +162,7 @@ import Foundation
     
     public func urlSession(_ session: URLSession, task: URLSessionTask, didCompleteWithError error: Error?) {
         
-        var fileName: String = "", serverUrl: String = "", etag: String?, ocId: String?, date: NSDate?, dateLastModified: NSDate?, length: Double = 0
+        var fileName: String = "", serverUrl: String = "", etag: String?, ocId: String?, date: NSDate?, dateLastModified: NSDate?, length: Int64 = 0
         let url = task.currentRequest?.url?.absoluteString.removingPercentEncoding
         if url != nil {
             fileName = (url! as NSString).lastPathComponent
@@ -207,7 +207,7 @@ import Foundation
             if let dateString = header["Last-Modified"] as? String {
                 dateLastModified = NCCommunicationCommon.shared.convertDate(dateString, format: "EEE, dd MMM y HH:mm:ss zzz")
             }
-            length = header["Content-Length"] as? Double ?? 0
+            length = header["Content-Length"] as? Int64 ?? 0
         }
         
         DispatchQueue.main.async {
